% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_tableau_size.R
\name{Adjust_Tableau_size}
\alias{Adjust_Tableau_size}
\title{Adjust_Tableau_size}
\usage{
Adjust_Tableau_size(import_files)
}
\arguments{
\item{import_files}{selected tableau file or folder, to change its size. If
it is only one file, make sure it is a twb file.}
}
\value{
tableau file with the correct size.
}
\description{
Adjust the size according to the 'stijlgids'in the selected tableau file/files.
}
