% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_tableau_colors.R
\name{Adjust_XML_color}
\alias{Adjust_XML_color}
\title{Adjust_XML_color}
\usage{
Adjust_XML_color(Variable_colors, Color_codes, Color_part, Value_list)
}
\arguments{
\item{Variable_colors}{Dataframe of common variables and their corresponding
colors.}

\item{Color_codes}{Dataframe of VU tableau colors and their color codes}

\item{Color_part}{part i  the XML file where the colors are}

\item{Value_list}{list of variables with a color in the XML file}
}
\description{
Adjust the colors in the XML file.
}
