% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barplot_per_level}
\alias{barplot_per_level}
\title{Barplot per level}
\usage{
barplot_per_level(x)
}
\arguments{
\item{x}{een categoriale vector}
}
\description{
Maakt een barplot per level van een factor.
}
