% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dashboard_names.R
\name{Check_dashboard_names}
\alias{Check_dashboard_names}
\title{Check_dashboard_names}
\usage{
Check_dashboard_names(import_files)
}
\arguments{
\item{import_files}{selected tableau file or folder,
to check the names of the dashboards.
If it is only one file, make sure it is a twb file.}
}
\value{
the wrongly named dashboards
}
\description{
Check the names of dashboards in tableau, according to the 'stijlgids'.
The wrong dashboards will be retuned together with the reason why it is a
wrong name
}
