% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_model_formula.R
\name{get_model_formula}
\alias{get_model_formula}
\title{Get model formula}
\usage{
get_model_formula(dependent, independent)
}
\arguments{
\item{dependent}{Specify dependent value.}

\item{independent}{Specify independent value.}
}
\value{
Formula for the model.
}
\description{
Creates a formula to use in a model.
}
