% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_model_variabelen}
\alias{plot_model_variabelen}
\title{Plot model variabelen.R}
\usage{
plot_model_variabelen(data, y, x, modelversie)
}
\arguments{
\item{data}{Dataframe}

\item{y}{Y-variabele van het model (string)}

\item{x}{Vector met X-variabelen (strings)}

\item{modelversie}{Stringtekst om te bepalen waar de plots worden weggeschreven}
}
\description{
Doel: De functie plot_model_variabelen maakt explorerende plots voor de
variabelen in het opgegeven model.
}
\details{
Afhankelijkheden: Index.R

Opmerkingen:
\enumerate{
\item Plots worden weggeschreven naar
/Datasets/Plots/Exploreren/(y)/(modelversie)/(Datum)
}
}
