% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotperformance_list}
\alias{plotperformance_list}
\title{Maakt aestetisch mooite plots in kleuren van VU}
\usage{
plotperformance_list(
  model_list,
  type_plot = "gain",
  noauc = F,
  legend = T,
  auc_ook_1 = F
)
}
\arguments{
\item{model_list}{Model lijst}

\item{type_plot}{Default: "Gain". Type plot}

\item{noauc}{Default: False. No AUC}

\item{legend}{Default: True. Legend.}

\item{auc_ook_1}{Default: False. AUC ook 1}
}
\value{
ggplot object, so you can save the resulting plot with ggsave:
ggsave(filename = "Gainchart.png", plot = plot.gain)
}
\description{
Maakt aestisch mooie plots van de volgende plots in kleuren VU: "ROC","ROC.ggplot","gain", "lift", "sens/spec","Precision/recall curve"
}
\examples{
\dontrun{plotperformance_list(model_list)}
}
