% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotperformance_single_by_var}
\alias{plotperformance_single_by_var}
\title{PlotPerformance_Single}
\usage{
plotperformance_single_by_var(
  predicted,
  real_values,
  type_plot = "gain",
  by = NULL,
  by_nam = "Periode",
  legend = T
)
}
\arguments{
\item{predicted}{Dataframe met de voorspelde probabilities voor de test-set}

\item{real_values}{De echte class voor de test set (for example, the class is "churner") with 0/1 values}

\item{type_plot}{Opties zijn: "ROC.ggplot", "ROC.colorize","gain" & "lift"}

\item{by}{bij}

\item{by_nam}{bij nam?}

\item{legend}{Legenda}
}
\description{
Maakt "ROC.ggplot", "ROC.colorize","gain" of "lift" plots in huiskleuren van de VU.
Voorbeelden aanroepen:
PlotPerformance(predicted.values, real.values, "ROC.colorize")
plot.ROC <- PlotPerformance(predicted.values, real.values, "ROC.ggplot")
plot.gain <- PlotPerformance(predicted.values, real.values, "gain")
plot.lift <- PlotPerformance(predicted.values, real.values, "lift")
}
\details{
De functie returns een ggplot object, die je op kunt slaan met ggsave:
#' ggsave(filename = "Gainchart.png", plot = plot.gain)
}
