% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots_van_model_homogeneity_normality}
\alias{plots_van_model_homogeneity_normality}
\title{Maakt plots over homogeneity en normality model}
\usage{
plots_van_model_homogeneity_normality(Y_variabele, X_variabelen, dfM)
}
\arguments{
\item{Y_variabele}{Y variabele}

\item{X_variabelen}{X Variabelen}

\item{dfM}{Dataframe}
}
\value{
None.
}
\description{
Maakt (en bewaart) verschillende plots ("B.1. Model") over homogeneity:
"B.1. Model Cooks distance", "B.1. Invloedrijke observaties", "B.2. Residuen vs gefitte waarden",
"B.2. Residual plot", "B.2. Boxplot Residuals" en normality: "C.1. Normality QQ",
"C.1. Normality QQ std residuals"
}
\examples{
\dontrun{plots_van_model_homogeneity_normality(Y_variabele, X_variabelen, dfM )}
}
