% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sa_boxplot}
\alias{sa_boxplot}
\title{SA boxplot}
\usage{
sa_boxplot(x, col = VU_BlueLight, ...)
}
\arguments{
\item{x}{Een vector of string}

\item{col}{Kleur om default VU kleur (VU_BlueLight) te overrulen.}

\item{...}{De overige parameters worden meegegeven aan de functie.}
}
\description{
Wrapper functie rondom boxplot uit package 'graphics'
}
