% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sa_category_plot}
\alias{sa_category_plot}
\title{SA Category Plot}
\usage{
sa_category_plot(data, x, y, title, remove_labels = 10)
}
\arguments{
\item{data}{De data frame}

\item{x}{Geef de waarde op voor x}

\item{y}{Geefde waarde op voor y}

\item{title}{Indien geen titel wordt opgegeven, automatisch titel
adv van 'x' en 'y'.}

\item{remove_labels}{Als het aantal categorieen hoger is dan remove_labels, wordt
het label verwijderd}
}
\description{
SA Category Plot
}
