% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sa_dotchart_hist}
\alias{sa_dotchart_hist}
\title{SA Dotchart Histogram}
\usage{
sa_dotchart_hist(data, var, title = NULL)
}
\arguments{
\item{data}{De data frame}

\item{var}{Een character string}

\item{title}{Indien geen titel wordt opgegeven, automatisch titel gegenereerd adv var naam}
}
\description{
Maakt een Dotchart en Histogram en voegt deze samen in een figuur.
}
