% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sa_dotchart_old}
\alias{sa_dotchart_old}
\title{SA dotchart old}
\usage{
sa_dotchart_old(x, lcolor = VU_BlueLight, ...)
}
\arguments{
\item{x}{Een string of vector.}

\item{lcolor}{Kleur om default VU kleur (VU_BlueLight) te overrulen.}

\item{...}{De overige parameters wroden meegegeven aan de functie.}
}
\description{
Wrapper functie rondom dotchart uit package 'graphics'
}
