% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{sa_drop_vif}
\alias{sa_drop_vif}
\title{SA drop vif}
\usage{
sa_drop_vif(df, threshold = 2.5)
}
\arguments{
\item{df}{The data frame}

\item{threshold}{The threshold, all variables with a VIF value above
this limit will be dropped.}
}
\value{
Data frame without correlated variables.
}
\description{
Function to drop variable based on their VIF value.
}
