% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{sa_line2user}
\alias{sa_line2user}
\title{SA line2user}
\usage{
sa_line2user(line, side)
}
\arguments{
\item{line}{Line}

\item{side}{A value of 1, 2, 3, or 4.}
}
\description{
Function to create a title.
}
