% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sa_violinplot}
\alias{sa_violinplot}
\title{SA violinplot}
\usage{
sa_violinplot(
  df,
  variable,
  fill_violin = VU_BlueMiddle,
  fill_boxplot = VU_BlueLight
)
}
\arguments{
\item{df}{De data frame}

\item{variable}{Een character string}

\item{fill_violin}{Kleur van violin staat standaard op "VU_BlueMiddle"}

\item{fill_boxplot}{Kleur van boxplot staat standaard op "VU_Bluelight"}
}
\description{
Wrapper functie rondom violinplot.
}
